prof-jd-2_0-env-hw5 после перевода на Collections
