+++ src/main/java/dev/ji/employeenewnew/rest/EmployeeController.java
@@ -1,19 +1,21 @@
 package dev.ji.employeenewnew.rest;
 
 import dev.ji.employeenewnew.model.Employee;
-import dev.ji.employeenewnew.service.EmployeeService;
+import dev.ji.employeenewnew.service.EmployeeServiceCollections;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.util.Set;
+
 @RestController
 @RequestMapping("/employee")
 public class EmployeeController {
 
-    private EmployeeService employeeService;
+    private EmployeeServiceCollections employeeService;
 
-    public EmployeeController(EmployeeService employeeService) {
+    public EmployeeController(EmployeeServiceCollections employeeService) {
         this.employeeService = employeeService;
     }
 
@@ -42,5 +44,11 @@ public class EmployeeController {
         Employee foundEmployee = employeeService.findEmployee(firstName, lastName);
         return foundEmployee;
     }
+
+    @GetMapping("/all")
+    public Set<Employee> getAllEmployees() {
+        return employeeService.getAllEmployees();
+    }
+
 
 }
