+++ src/main/java/dev/ji/employeenewnew/service/EmployeeServiceArchive.java
@@ -1,0 +1,15 @@
+//package dev.ji.employeenewnew.service;
+//
+//import dev.ji.employeenewnew.model.Employee;
+//
+//import java.util.Set;
+//
+//public interface EmployeeServiceArchive {
+//    String addEmployee(String firstName, String lastName);
+//
+//    String removeEmployee(String firstName, String lastName);
+//
+//    Employee findEmployee(String firstName, String lastName);
+//
+//    Set<Employee> getAllEmployees();
+//}
